feat(runtime): depth-aware dim mask (Task 5) + bundled inline-text-flow WIP
My Task 5 work: introduce UI_DIM_MASK (0x7F7F7F under UI_COLOR_DEPTH 888,
0x7BEF under 565) and convert the three dimFg locals to UI_COLOR_T so 888
dimming halves each 8-bit channel independently. Under 565 the mask and
type resolve to the prior values.

Also commits pre-existing uncommitted WIP from the inline-text-flow effort
that was present in the working tree at the start of this session: UINode
padding fields, ui_mark_scroll_view_dirty scroll-canvas helpers, radio-group
index uint16_t widening, showcase.ui expansion, and associated preview/test
updates. All 133 tests across the affected suites pass; cuttlefish builds clean.

+++ packages/cuttlefish/src/ir/ui-element-auto-wire.ts
@@ -220,7 +220,7 @@ function autoWireNode(treeName: string, node: AutoWireNode, nodeIndex: number): 
       nodeIndex,
       kind: "click",
       fnName: `__ui_${node.id}_autoclick`,
-      callbackBody: `for (uint8_t __r = 0; __r < __ui_radio_groups[${groupIdx}].count; __r++) { uint8_t __rn = __ui_radio_groups[${groupIdx}].nodeIndices[__r]; __ui_nodes[__rn].value = 0; ui_mark_dirty(__rn); } __ui_nodes[${nodeIndex}].value = 1; ui_mark_dirty(${nodeIndex});`,
+      callbackBody: `for (uint8_t __r = 0; __r < __ui_radio_groups[${groupIdx}].count; __r++) { uint16_t __rn = __ui_radio_groups[${groupIdx}].nodeIndices[__r]; __ui_nodes[__rn].value = 0; ui_mark_dirty(__rn); } __ui_nodes[${nodeIndex}].value = 1; ui_mark_dirty(${nodeIndex});`,
     });
   }
